feat: by_ and with_ methods supports types using constraints instead of a fixed reference based type.
