+++ lorm-macros/src/orm/mod.rs
@@ -4,8 +4,9 @@ mod save;
 mod select;
 mod with;
 
-use crate::helpers::*;
 use crate::models::OrmModel;
+use crate::utils::executor_type;
+use crate::utils::*;
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::punctuated::Punctuated;
@@ -53,9 +54,10 @@ pub fn expand_derive_to_orm_struct(
 ) -> syn::Result<TokenStream> {
     let model = OrmModel::from_fields(input, fields)?;
     let executor_type = executor_type(input)?;
+    let database_type = database_type(input)?;
 
-    let with_code = with::generate_with(&executor_type, &model)?;
-    let by_code = by::generate_by(&executor_type, &model)?;
+    let with_code = with::generate_with(&executor_type, &database_type, &model)?;
+    let by_code = by::generate_by(&executor_type, &database_type, &model)?;
     let select_code = select::generate_select(&executor_type, &model)?;
     let delete_code = delete::generate_delete(&executor_type, &model)?;
     let save_code = save::generate_save(&executor_type, &model)?;
