+++ lorm/Cargo.toml
@@ -20,7 +20,7 @@ postgres = ["lorm-macros/postgres", "sqlx/postgres"]
 mysql = ["lorm-macros/mysql", "sqlx/mysql"]
 
 [dependencies]
-lorm-macros = { path = "../lorm-macros", version = "0.0.7", optional = true }
+lorm-macros = { path = "../lorm-macros", optional = true }
 sqlx = { workspace = true, features = ["derive"] }
 serde = { workspace = true, features = ["std", "derive"] }
 thiserror = { workspace = true, features = ["std"] }
